+++ src/lib/feature/posts/ui/post/postComment/PostComment.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import { Typography } from '@/components/ui/typography/Typography'
-import { formatTimeAgo } from '@/utils/formatTimeAgo'
+import { formatTimeAgo } from '@/utils'
 import { Comment } from '@/lib/feature/posts/api/postsApi.types'
 import { useLazyGetCommentAnswerQuery } from '@/lib/feature/posts/api/postsApi'
 import PostAnswer from '@/lib/feature/posts/ui/post/postComment/postAnswer/PostAnswer'
