+++ src/utils/formatDate.ts
@@ -6,3 +6,11 @@ export function formatDate(dateString: string): string {
     month: 'long',
   })
 }
+
+export const formatDateToDotFormat = (isoDate: string): string => {
+  const date = new Date(isoDate)
+  const day = String(date.getDate()).padStart(2, '0')
+  const month = String(date.getMonth() + 1).padStart(2, '0') // месяцы с 0
+  const year = date.getFullYear()
+  return `${day}.${month}.${year}`
+}
